@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.1
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: adhan_service.proto
 
@@ -442,7 +442,7 @@ func file_adhan_service_proto_rawDescGZIP() []byte {
 }
 
 var file_adhan_service_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_adhan_service_proto_goTypes = []interface{}{
+var file_adhan_service_proto_goTypes = []any{
 	(*AdhanFile)(nil),               // 0: limestone.AdhanFile
 	(*CreateAdhanFileRequest)(nil),  // 1: limestone.CreateAdhanFileRequest
 	(*UpdateAdhanFileRequest)(nil),  // 2: limestone.UpdateAdhanFileRequest
@@ -477,7 +477,7 @@ func file_adhan_service_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_adhan_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_adhan_service_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*AdhanFile); i {
 			case 0:
 				return &v.state
@@ -489,7 +489,7 @@ func file_adhan_service_proto_init() {
 				return nil
 			}
 		}
-		file_adhan_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_adhan_service_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateAdhanFileRequest); i {
 			case 0:
 				return &v.state
@@ -501,7 +501,7 @@ func file_adhan_service_proto_init() {
 				return nil
 			}
 		}
-		file_adhan_service_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_adhan_service_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateAdhanFileRequest); i {
 			case 0:
 				return &v.state
@@ -513,7 +513,7 @@ func file_adhan_service_proto_init() {
 				return nil
 			}
 		}
-		file_adhan_service_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_adhan_service_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetAdhanFileRequest); i {
 			case 0:
 				return &v.state
@@ -525,7 +525,7 @@ func file_adhan_service_proto_init() {
 				return nil
 			}
 		}
-		file_adhan_service_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_adhan_service_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteAdhanFileRequest); i {
 			case 0:
 				return &v.state
@@ -537,7 +537,7 @@ func file_adhan_service_proto_init() {
 				return nil
 			}
 		}
-		file_adhan_service_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_adhan_service_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteAdhanFileResponse); i {
 			case 0:
 				return &v.state
