@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.1
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: event_service.proto
 
@@ -808,7 +808,7 @@ func file_event_service_proto_rawDescGZIP() []byte {
 
 var file_event_service_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
 var file_event_service_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_event_service_proto_goTypes = []interface{}{
+var file_event_service_proto_goTypes = []any{
 	(Event_GenderRestriction)(0),  // 0: limestone.Event.GenderRestriction
 	(Event_EventType)(0),          // 1: limestone.Event.EventType
 	(*Event)(nil),                 // 2: limestone.Event
@@ -854,7 +854,7 @@ func file_event_service_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_event_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Event); i {
 			case 0:
 				return &v.state
@@ -866,7 +866,7 @@ func file_event_service_proto_init() {
 				return nil
 			}
 		}
-		file_event_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateEventRequest); i {
 			case 0:
 				return &v.state
@@ -878,7 +878,7 @@ func file_event_service_proto_init() {
 				return nil
 			}
 		}
-		file_event_service_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateEventRequest); i {
 			case 0:
 				return &v.state
@@ -890,7 +890,7 @@ func file_event_service_proto_init() {
 				return nil
 			}
 		}
-		file_event_service_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteEventRequest); i {
 			case 0:
 				return &v.state
@@ -902,7 +902,7 @@ func file_event_service_proto_init() {
 				return nil
 			}
 		}
-		file_event_service_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteEventResponse); i {
 			case 0:
 				return &v.state
@@ -914,7 +914,7 @@ func file_event_service_proto_init() {
 				return nil
 			}
 		}
-		file_event_service_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*GetEventRequest); i {
 			case 0:
 				return &v.state
@@ -926,7 +926,7 @@ func file_event_service_proto_init() {
 				return nil
 			}
 		}
-		file_event_service_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*ListEventsRequest); i {
 			case 0:
 				return &v.state
@@ -938,7 +938,7 @@ func file_event_service_proto_init() {
 				return nil
 			}
 		}
-		file_event_service_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_event_service_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*ListEventsResponse); i {
 			case 0:
 				return &v.state
