@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: adhan_service.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	AdhanService_CreateAdhanFile_FullMethodName = "/limestone.AdhanService/CreateAdhanFile"
@@ -48,8 +48,9 @@ func NewAdhanServiceClient(cc grpc.ClientConnInterface) AdhanServiceClient {
 }
 
 func (c *adhanServiceClient) CreateAdhanFile(ctx context.Context, in *CreateAdhanFileRequest, opts ...grpc.CallOption) (*AdhanFile, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AdhanFile)
-	err := c.cc.Invoke(ctx, AdhanService_CreateAdhanFile_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AdhanService_CreateAdhanFile_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -57,8 +58,9 @@ func (c *adhanServiceClient) CreateAdhanFile(ctx context.Context, in *CreateAdha
 }
 
 func (c *adhanServiceClient) UpdateAdhanFile(ctx context.Context, in *UpdateAdhanFileRequest, opts ...grpc.CallOption) (*AdhanFile, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AdhanFile)
-	err := c.cc.Invoke(ctx, AdhanService_UpdateAdhanFile_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AdhanService_UpdateAdhanFile_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -66,8 +68,9 @@ func (c *adhanServiceClient) UpdateAdhanFile(ctx context.Context, in *UpdateAdha
 }
 
 func (c *adhanServiceClient) GetAdhanFile(ctx context.Context, in *GetAdhanFileRequest, opts ...grpc.CallOption) (*AdhanFile, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AdhanFile)
-	err := c.cc.Invoke(ctx, AdhanService_GetAdhanFile_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AdhanService_GetAdhanFile_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -75,8 +78,9 @@ func (c *adhanServiceClient) GetAdhanFile(ctx context.Context, in *GetAdhanFileR
 }
 
 func (c *adhanServiceClient) DeleteAdhanFile(ctx context.Context, in *DeleteAdhanFileRequest, opts ...grpc.CallOption) (*DeleteAdhanFileResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteAdhanFileResponse)
-	err := c.cc.Invoke(ctx, AdhanService_DeleteAdhanFile_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, AdhanService_DeleteAdhanFile_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -85,7 +89,7 @@ func (c *adhanServiceClient) DeleteAdhanFile(ctx context.Context, in *DeleteAdha
 
 // AdhanServiceServer is the server API for AdhanService service.
 // All implementations must embed UnimplementedAdhanServiceServer
-// for forward compatibility
+// for forward compatibility.
 type AdhanServiceServer interface {
 	// Creates an adhan audio file in the database for a masjid.
 	CreateAdhanFile(context.Context, *CreateAdhanFileRequest) (*AdhanFile, error)
@@ -98,9 +102,12 @@ type AdhanServiceServer interface {
 	mustEmbedUnimplementedAdhanServiceServer()
 }
 
-// UnimplementedAdhanServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedAdhanServiceServer struct {
-}
+// UnimplementedAdhanServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedAdhanServiceServer struct{}
 
 func (UnimplementedAdhanServiceServer) CreateAdhanFile(context.Context, *CreateAdhanFileRequest) (*AdhanFile, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateAdhanFile not implemented")
@@ -115,6 +122,7 @@ func (UnimplementedAdhanServiceServer) DeleteAdhanFile(context.Context, *DeleteA
 	return nil, status.Errorf(codes.Unimplemented, "method DeleteAdhanFile not implemented")
 }
 func (UnimplementedAdhanServiceServer) mustEmbedUnimplementedAdhanServiceServer() {}
+func (UnimplementedAdhanServiceServer) testEmbeddedByValue()                      {}
 
 // UnsafeAdhanServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to AdhanServiceServer will
@@ -124,6 +132,13 @@ type UnsafeAdhanServiceServer interface {
 }
 
 func RegisterAdhanServiceServer(s grpc.ServiceRegistrar, srv AdhanServiceServer) {
+	// If the following call pancis, it indicates UnimplementedAdhanServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&AdhanService_ServiceDesc, srv)
 }
 
