@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: masjid_service.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	MasjidService_CreateMasjid_FullMethodName = "/limestone.MasjidService/CreateMasjid"
@@ -52,8 +52,9 @@ func NewMasjidServiceClient(cc grpc.ClientConnInterface) MasjidServiceClient {
 }
 
 func (c *masjidServiceClient) CreateMasjid(ctx context.Context, in *CreateMasjidRequest, opts ...grpc.CallOption) (*Masjid, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Masjid)
-	err := c.cc.Invoke(ctx, MasjidService_CreateMasjid_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MasjidService_CreateMasjid_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -61,8 +62,9 @@ func (c *masjidServiceClient) CreateMasjid(ctx context.Context, in *CreateMasjid
 }
 
 func (c *masjidServiceClient) UpdateMasjid(ctx context.Context, in *UpdateMasjidRequest, opts ...grpc.CallOption) (*Masjid, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Masjid)
-	err := c.cc.Invoke(ctx, MasjidService_UpdateMasjid_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MasjidService_UpdateMasjid_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -70,8 +72,9 @@ func (c *masjidServiceClient) UpdateMasjid(ctx context.Context, in *UpdateMasjid
 }
 
 func (c *masjidServiceClient) DeleteMasjid(ctx context.Context, in *DeleteMasjidRequest, opts ...grpc.CallOption) (*DeleteMasjidResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteMasjidResponse)
-	err := c.cc.Invoke(ctx, MasjidService_DeleteMasjid_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MasjidService_DeleteMasjid_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -79,8 +82,9 @@ func (c *masjidServiceClient) DeleteMasjid(ctx context.Context, in *DeleteMasjid
 }
 
 func (c *masjidServiceClient) GetMasjid(ctx context.Context, in *GetMasjidRequest, opts ...grpc.CallOption) (*Masjid, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Masjid)
-	err := c.cc.Invoke(ctx, MasjidService_GetMasjid_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MasjidService_GetMasjid_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -88,8 +92,9 @@ func (c *masjidServiceClient) GetMasjid(ctx context.Context, in *GetMasjidReques
 }
 
 func (c *masjidServiceClient) ListMasjids(ctx context.Context, in *ListMasjidsRequest, opts ...grpc.CallOption) (*ListMasjidsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListMasjidsResponse)
-	err := c.cc.Invoke(ctx, MasjidService_ListMasjids_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MasjidService_ListMasjids_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -98,7 +103,7 @@ func (c *masjidServiceClient) ListMasjids(ctx context.Context, in *ListMasjidsRe
 
 // MasjidServiceServer is the server API for MasjidService service.
 // All implementations must embed UnimplementedMasjidServiceServer
-// for forward compatibility
+// for forward compatibility.
 type MasjidServiceServer interface {
 	// Creates a masjid in the database.
 	CreateMasjid(context.Context, *CreateMasjidRequest) (*Masjid, error)
@@ -114,9 +119,12 @@ type MasjidServiceServer interface {
 	mustEmbedUnimplementedMasjidServiceServer()
 }
 
-// UnimplementedMasjidServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedMasjidServiceServer struct {
-}
+// UnimplementedMasjidServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMasjidServiceServer struct{}
 
 func (UnimplementedMasjidServiceServer) CreateMasjid(context.Context, *CreateMasjidRequest) (*Masjid, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateMasjid not implemented")
@@ -134,6 +142,7 @@ func (UnimplementedMasjidServiceServer) ListMasjids(context.Context, *ListMasjid
 	return nil, status.Errorf(codes.Unimplemented, "method ListMasjids not implemented")
 }
 func (UnimplementedMasjidServiceServer) mustEmbedUnimplementedMasjidServiceServer() {}
+func (UnimplementedMasjidServiceServer) testEmbeddedByValue()                       {}
 
 // UnsafeMasjidServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MasjidServiceServer will
@@ -143,6 +152,13 @@ type UnsafeMasjidServiceServer interface {
 }
 
 func RegisterMasjidServiceServer(s grpc.ServiceRegistrar, srv MasjidServiceServer) {
+	// If the following call pancis, it indicates UnimplementedMasjidServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MasjidService_ServiceDesc, srv)
 }
 
